pyproject.toml
[project]
name = "unstructured-take-home"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.97.0",
    "chromadb>=1.5.8",
    "dspy-ai>=3.2.0",
    "huggingface-hub>=1.11.0",
    "ipykernel>=7.2.0",
    "langgraph>=1.1.9",
    "onnxruntime<1.21",
    "pandas>=3.0.2",
    "pillow>=12.2.0",
    "pyarrow>=24.0.0",
    "rarfile>=4.2",
    "tiktoken>=0.12.0",
]

[dependency-groups]
dev = [
    "ruff>=0.15.11",
]
